requirements.txt
numpy==2.3.1
httpx==0.28.1
python-dotenv==1.1.1
sentence-transformers==5.0.0
